app/services/analyzer.py
import os
import time
import random
from pathlib import Path
from typing import Dict, Tuple, List
import logging

from PIL import Image
import numpy as np

from app.models.plant_analysis import PlantAnalysisResponse, PlantStatus, AnalysisResult
from app.config import get_settings

settings = get_settings()

# Configure logger
logger = logging.getLogger(__name__)
logger.setLevel(logging.INFO)


# Placeholder class for the ML model
class PlantAnalysisModel:
    """
    Placeholder for the Machine Learning model for plant analysis.

    This class would be replaced by a real ML model in a production implementation,
    such as a trained CNN, ResNet, etc.
    """

    def __init__(self, model_path: str):
        """
        Initialize the model.

        Args:
            model_path: Path to the saved model
        """
        self.model_path = model_path
        self.loaded = False

        # Here the actual model would be loaded
        # e.g., with TensorFlow or PyTorch
        # self.model = tf.keras.models.load_model(model_path)
        # or
        # self.model = torch.load(model_path)

        logger.info(f"Model placeholder initialized. Would load from: {model_path}")
        self.loaded = True

        # Predefined possible results for our placeholder
        self.possible_issues = [
            {
                "status": PlantStatus.UNHEALTHY,
                "details": "Yellowing leaves detected, possible nutrient deficiency (Nitrogen).",
                "recommendations": "Apply a balanced liquid fertilizer rich in Nitrogen. Follow product instructions carefully."
            },
            {
                "status": PlantStatus.UNHEALTHY,
                "details": "Signs of pest damage (aphids) on lower leaves.",
                "recommendations": "Spray the plant with insecticidal soap, focusing on the undersides of leaves. Repeat application if necessary."
            },
            {
                "status": PlantStatus.UNHEALTHY,
                "details": "Dark spots on leaves, potential early blight.",
                "recommendations": "Remove and destroy affected leaves immediately. Ensure good air circulation around the plant. Consider applying a fungicide if the problem persists."
            },
            {
                "status": PlantStatus.UNHEALTHY,
                "details": "Wilting observed, check soil moisture levels.",
                "recommendations": "Water the plant thoroughly if the soil is dry. Ensure proper drainage to prevent overwatering."
            },
            {
                "status": PlantStatus.HEALTHY,
                "details": "Plant appears healthy. No significant issues detected.",
                "recommendations": "Continue with regular care and monitoring."
            }
        ]

    def preprocess_image(self, image: Image.Image) -> np.ndarray:
        """
        Preprocess the image for the model.

        Args:
            image: PIL Image object

        Returns:
            Preprocessed image as numpy array
        """
        # Example preprocessing
        # 1. Resize
        resized_img = image.resize((224, 224))

        # 2. Convert to array and normalize
        img_array = np.array(resized_img)

        # 3. Additional processing depending on the model
        # e.g., normalization, etc.

        return img_array

    def predict(self, image: Image.Image) -> Tuple[Dict, float]:
        """
        Make a prediction on the image.

        Args:
            image: PIL Image object

        Returns:
            Tuple of prediction result and confidence score
        """
        # Here the actual prediction would take place
        # preprocessed = self.preprocess_image(image)
        # prediction = self.model.predict(preprocessed)

        # For the placeholder: Random results
        if random.random() > 0.4:  # 60% probability for healthy
            result = self.possible_issues[4]  # Healthy
            confidence = random.uniform(0.7, 0.98)
        else:
            # Select a random problem
            result = random.choice(self.possible_issues[:4])  # One of the problems
            confidence = random.uniform(0.65, 0.95)

        return result, confidence


class PlantAnalyzerService:
    """Service for analyzing plant images"""

    def __init__(self):
        """Initialize the Analyzer Service"""
        model_path = settings.MODEL_PATH
        self.model = PlantAnalysisModel(model_path)

        # Create directory for upload files if it doesn't exist
        os.makedirs(settings.UPLOAD_DIR, exist_ok=True)

    async def analyze_image(self, image_path: str) -> AnalysisResult:
        """
        Analyze a plant image and return the results.

        Args:
            image_path: Path to the saved image

        Returns:
            AnalysisResult with the results of the analysis
        """
        start_time = time.time()

        try:
            # Load image
            img = Image.open(image_path)
            width, height = img.size

            # Perform analysis
            prediction, confidence = self.model.predict(img)

            # Create result
            plant_analysis = PlantAnalysisResponse(
                status=prediction["status"],
                details=prediction["details"],
                recommendations=prediction.get("recommendations")
            )

            # Calculate processing time
            processing_time = int((time.time() - start_time) * 1000)

            return AnalysisResult(
                plant_analysis=plant_analysis,
                confidence_score=confidence,
                processing_time_ms=processing_time,
                image_width=width,
                image_height=height
            )

        except Exception as e:
            logger.error(f"Error during image analysis: {str(e)}")
            raise RuntimeError(f"Error during image analysis: {str(e)}")


# Create singleton instance of the service
analyzer_service = PlantAnalyzerService()