requirements.txt
annotated-types==0.7.0
anyio==4.9.0
fastapi==0.115.12
idna==3.10
numpy==2.2.5
pillow==11.2.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
